requirements.txt
pandas
numpy
matplotlib
plotly
scikit-learn
xgboost
lightgbm
catboost
optuna
torch
pyyaml
ipywidgets
mlflow

